HUD reconstruct
